refactor(types): retype ReliabilityReport.Score as float64 0.0–1.0
The persisted Score field was an int percent (0–100), which forced
every consumer to guess whether "67" meant percent or fraction.
Switch to the canonical fractional form; renderers multiply by 100
when displaying.

- pkg/types/report.go: Score int → float64
- cmd/ruptor/main.go: compute as float; pass int(*100) to UI completion
- internal/report/stdout_renderer.go: render with %.0f%% via *100
- internal/report/html_renderer.go: register pct FuncMap helper
- web/chaos_report.html.tmpl: use {{pct .Score}} in display card
- internal/report/stdout_renderer_test.go: assertion + sample retype

No behavior change at the user surface — "Score: 67%" still renders
identically. Suite 259 passing.

+++ pkg/types/report.go
@@ -9,15 +9,19 @@ import "time"
 const ReportSchemaVersion = "1.0"
 
 type ReliabilityReport struct {
-	SchemaVersion string       `json:"schema_version"`
-	RuptorVersion string       `json:"ruptor_version"`
-	AgentName     string       `json:"agent_name"`
-	RunAt         time.Time    `json:"run_at"`
-	TotalTests    int          `json:"total_tests"`
-	Passed        int          `json:"passed"`
-	Failed        int          `json:"failed"`
-	Score         int          `json:"score"`
-	Results       []TestResult `json:"results"`
+	SchemaVersion string    `json:"schema_version"`
+	RuptorVersion string    `json:"ruptor_version"`
+	AgentName     string    `json:"agent_name"`
+	RunAt         time.Time `json:"run_at"`
+	TotalTests    int       `json:"total_tests"`
+	Passed        int       `json:"passed"`
+	Failed        int       `json:"failed"`
+	// Score is the Robustness Score in the canonical 0.0–1.0 range
+	// (1.0 = every test passed). Renderers multiply by 100 to show
+	// "67%". Persisted in this form so platform-side analytics never
+	// have to guess between "67" (percent) and "0.67" (fraction).
+	Score   float64      `json:"score"`
+	Results []TestResult `json:"results"`
 }
 
 type ConversationReport struct {
